feat(MissionsUtils): Moving static methods into Utils
